infra/scripts/requirements.txt
azure-identity==1.19.0
python-dotenv==1.0.1
azure-search-documents==11.6.0b4
azure-cosmos==4.7.0
azure-storage-blob
rich
invoke==2.2.0
setuptools>=75.8.0
cffi>=1.17.1
cryptography>=44.0.0
azure.mgmt.eventgrid>=10.2.0